Create networkDataSource + localDataSource + relocate their respective files
