+++ app/src/main/kotlin/com/architects/pokearch/core/data/local/LocalDataSource.kt
@@ -1,0 +1,4 @@
+package com.architects.pokearch.core.data.local
+
+class LocalDataSource {
+}
