+++ app/src/main/kotlin/com/architects/pokearch/core/data/local/database/dao/PokemonInfoDao.kt
@@ -1,10 +1,10 @@
-package com.architects.pokearch.core.data.database.dao
+package com.architects.pokearch.core.data.local.database.dao
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.architects.pokearch.core.data.database.entities.PokemonInfoEntity
+import com.architects.pokearch.core.data.local.database.entities.PokemonInfoEntity
 
 @Dao
 interface PokemonInfoDao {
