+++ app/src/main/kotlin/com/architects/pokearch/core/data/local/database/entities/converters/TypesHolderConverter.kt
@@ -1,4 +1,4 @@
-package com.architects.pokearch.core.data.database.entities.converters
+package com.architects.pokearch.core.data.local.database.entities.converters
 
 import androidx.room.TypeConverter
 import com.google.gson.Gson
