+++ app/src/main/kotlin/com/architects/pokearch/core/data/network/service/PokedexService.kt
@@ -1,7 +1,7 @@
 package com.architects.pokearch.core.data.network.service
 
-import com.architects.pokearch.core.data.model.NetworkPokemons
-import com.architects.pokearch.core.data.model.NetworkPokemonInfo
+import com.architects.pokearch.core.data.network.model.NetworkPokemons
+import com.architects.pokearch.core.data.network.model.NetworkPokemonInfo
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Path
