+++ app/src/main/kotlin/com/architects/pokearch/core/data/repository/PokeArchRepository.kt
@@ -2,12 +2,12 @@ package com.architects.pokearch.core.data.repository
 
 import android.util.Log
 import arrow.core.Either
-import com.architects.pokearch.core.data.database.dao.PokemonDao
-import com.architects.pokearch.core.data.database.dao.PokemonInfoDao
+import com.architects.pokearch.core.data.local.database.dao.PokemonDao
+import com.architects.pokearch.core.data.local.database.dao.PokemonInfoDao
 import com.architects.pokearch.core.data.mappers.PokemonEntityMapper
 import com.architects.pokearch.core.data.mappers.PokemonInfoEntityMapper
+import com.architects.pokearch.core.data.network.PokedexNetwortkDataSource
 import com.architects.pokearch.core.data.network.service.CryService
-import com.architects.pokearch.core.data.network.service.PokedexService
 import com.architects.pokearch.core.domain.model.Failure
 import com.architects.pokearch.core.domain.model.Pokemon
 import com.architects.pokearch.core.domain.model.PokemonInfo
@@ -17,7 +17,7 @@ import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
 class PokeArchRepository @Inject constructor(
-    private val pokedexService: PokedexService,
+    private val pokedexNetworkDataSource: PokedexNetwortkDataSource,
     private val cryService: CryService,
     private val pokemonDao: PokemonDao,
     private val pokemonInfoDao: PokemonInfoDao,
@@ -47,47 +47,12 @@ class PokeArchRepository @Inject constructor(
             )
         )
 
-        if (areMorePokemonAvailableRemote()) {
-            emit(getRemotePokemonList(filter, limit, offset))
+        if (pokedexNetworkDataSource.areMorePokemonAvailableRemote(pokemonDao.countPokemonList())) {
+            emit(pokedexNetworkDataSource.getRemotePokemonList(filter, limit, offset))
         }
     }
 
-    private suspend fun areMorePokemonAvailableRemote() =
-        pokedexService.fetchPokemonList(1, pokemonDao.countPokemonList()).let { responseCount ->
-            when {
-                responseCount.isSuccessful -> {
-                    responseCount.body()?.next != null
-                }
 
-                else -> false
-            }
-        }
-
-    private suspend fun getRemotePokemonList(
-        filter: String,
-        limit: Int,
-        offset: Int,
-    ) =
-        pokedexService.fetchPokemonList(LIMIT_ALL, 0).let { response ->
-            when {
-                response.isSuccessful -> {
-                    response.body()?.let { pokemonResponse ->
-                        pokemonDao.insertPokemonList(PokemonEntityMapper.asEntity(pokemonResponse.results))
-                        Either.Right(
-                            PokemonEntityMapper.asDomain(
-                                pokemonDao.getPokemonList(
-                                    filter,
-                                    limit,
-                                    offset
-                                )
-                            )
-                        )
-                    } ?: Either.Left(Failure.UnknownError)
-                }
-
-                else -> Either.Left(Failure.UnknownError)
-            }
-        }
 
 
     override suspend fun fetchPokemonInfo(id: Int): Flow<Either<Failure, PokemonInfo>> = flow {
@@ -96,29 +61,9 @@ class PokeArchRepository @Inject constructor(
         }
 
         if (pokemon == null) {
-            emit(getRemotePokemon(id))
+            emit(pokedexNetworkDataSource.getRemotePokemon(id))
         }
     }
-
-    private suspend fun getRemotePokemon(
-        id: Int,
-    ): Either<Failure, PokemonInfo> {
-        val response = pokedexService.fetchPokemonInfo(id)
-
-        return when {
-            response.isSuccessful -> {
-                response.body()?.let {
-                    pokemonInfoDao.insertPokemonInfo(PokemonInfoEntityMapper.asEntity(it))
-                    pokemonInfoDao.getPokemonInfo(id)?.let { entity ->
-                        Either.Right(PokemonInfoEntityMapper.asDomain(entity))
-                    }
-                } ?: Either.Left(Failure.UnknownError)
-            }
-
-            else -> Either.Left(Failure.UnknownError)
-        }
-    }
-
 
     override suspend fun fetchCry(name: String): String {
         var result = ""
